requirements.txt
Flask==1.1.1
Jinja2==2.10.1
gunicorn==19.9.0
certifi==2019.11.28
itsdangerous==1.1.0
tensorflow-cpu
scikit-image
numpy
werkzeug
